+++ src/utils/NodeUtils.ts
@@ -1,7 +1,7 @@
 import NodeType from '~/enums/nodeType';
 import { INodeBase } from '~/models/INode';
 import NodeLocationType from '~/types/NodeLocationType';
-import * as s from './nodeTypeColors.scss';
+import * as s from './nodeUtils.scss';
 
 class NodeUtils {
     public static getNodeTypeClass = (
