+++ src/utils/transitTypeUtils.scss
